Rework User module on CRUD repository
