+++ packages/server-scala/src/main/scala/modules/user/guice/modules/UserModule.scala
@@ -1,12 +1,19 @@
 package modules.user.guice.modules
 
+import com.byteslounge.slickrepo.repository.Repository
+import com.google.inject.Provides
 import modules.user.graphql.resolvers.{UserResolver, UserResolverImpl}
-import modules.user.repositories.{UserRepo, UserRepoImpl}
+import modules.user.model.User
+import modules.user.repositories.UserRepository
 import net.codingwell.scalaguice.ScalaModule
+import slick.jdbc.JdbcProfile
 
 class UserModule extends ScalaModule {
-  override def configure() = {
+
+  override def configure() {
     bind[UserResolver].to[UserResolverImpl]
-    bind[UserRepo].to[UserRepoImpl]
   }
+
+  @Provides
+  def userRepository(driver: JdbcProfile): Repository[User, Int] = new UserRepository(driver)
 }
