+++ packages/server-scala/src/main/scala/modules/user/repositories/UserSchemaInitializer.scala
@@ -2,7 +2,8 @@ package modules.user.repositories
 
 import core.slick.{SchemaInitializer, SchemaUtil}
 import javax.inject.Inject
-import modules.user.model.User
+import modules.user.model.UserTable
+import modules.user.model.UserTable.UserTable
 import slick.jdbc.SQLiteProfile.api._
 import slick.lifted.TableQuery
 
@@ -12,16 +13,16 @@ class UserSchemaInitializer @Inject()(database: Database)
                                      (implicit executionContext: ExecutionContext) extends SchemaInitializer
   with SchemaUtil {
 
-  val users: TableQuery[User.Table] = TableQuery[User.Table]
+  val users: TableQuery[UserTable] = TableQuery[UserTable]
 
   override def create(): Future[Unit] = {
-    withTable(database, users, User.name, _.isEmpty) {
+    withTable(database, users, UserTable.name, _.isEmpty) {
       DBIO.seq(users.schema.create)
     }
   }
 
   override def drop(): Future[Unit] = {
-    withTable(database, users, User.name, _.nonEmpty) {
+    withTable(database, users, UserTable.name, _.nonEmpty) {
       DBIO.seq(users.schema.drop)
     }
   }
